+++ src/app.js
@@ -6,7 +6,7 @@ const passport = require('passport');
 const { Strategy, ExtractJwt } = require('passport-jwt');
 const { Logger } = require('./tools/Logger');
 const config = require('./config/config');
-const errorHandler = require('./tools/ErrorHandler');
+const errorHandler = require('./middlewares/ErrorHandler');
 const userRouter = require('./resources/user/user.router');
 const transactionRouter = require('./resources/transaction/transaction.router');
 const accountRouter = require('./resources/account/account.router');
@@ -17,7 +17,7 @@ const authRouter = require('./resources/auth/auth.router');
 const faqRouter = require('./resources/faq/faq.router');
 const piggyBankRouter = require('./resources/piggy_bank/piggy.router');
 const { jwtCallback } = require('./resources/auth/auth.repository');
-const addGuard = require('./tools/guards');
+const addGuard = require('./middlewares/guards');
 const db = require('./loader/dbconnect');
 
 db.connect();
@@ -43,7 +43,7 @@ app.use('/account', auth, accountRouter);
 app.use('/statistic', auth, addGuard, statisticRouter);
 app.use('/category', auth, categoryRouter);
 app.use('/piggybank', auth, piggyBankRouter);
-app.use('/subs', auth, subscriptionRouter);
+app.use('/subscription', auth, subscriptionRouter);
 app.use('/auth', authRouter);
 app.use('/faq', faqRouter);
 app.use('/', require('./resources/home.router'));
